меняет структуру проекта, добавляет unit tests
