+++ index.php
@@ -2,8 +2,8 @@
 $user_name = 'Михаил';
 $title = 'Главная страница';
 
-require_once('data.php');
-require_once('helpers.php');
+require_once __DIR__ .'/data.php';
+require_once __DIR__ .'/bootstrap.php';
 
 $content = include_template('main.php', ['lots' => $lots, 'categories'=> $categories] );
 $layout_content = include_template('layout.php',['content' => $content, 'user_name' => $user_name,
